% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BFDR}
\alias{BFDR}
\title{BFDR}
\usage{
BFDR(ppa)
}
\arguments{
\item{ppa}{the vector of posterior probabilities}
}
\description{
This function, given a vector of posterior probabilities,
computes the Bayesian False Discovery Rate
}
\examples{
BFDR()
}
\keyword{BFDR}

