% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BFDR_optimal}
\alias{BFDR_optimal}
\title{Find optimal choice of kappa based on EMBFDR}
\usage{
BFDR_optimal(PPA_MAT, KAPPA_VEC, bfdr_threshold, prt_threshold)
}
\arguments{
\item{PPA_MAT}{the matrix of conditional posterior probabilities}

\item{KAPPA_VEC}{the vector of kappa values used in the fitting}

\item{bfdr_threshold}{the maximum desired EMBFDR}

\item{prt_threshold}{the maximum desired percent of markers selected}
}
\description{
This function, given a matrix of posterior probabilities (PPA_MAT)
obtained when using a variety of kappa values (KAPPA_VEC), finds
the optimal choice of kappa based on our BFDR criterion
}
\examples{
BFDR_optimal()
}
\keyword{EMBFDR}

