% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculate_spatial_boost}
\alias{calculate_spatial_boost}
\title{Spatial Boost Term Calculation}
\usage{
calculate_spatial_boost(SX, CHRS = NULL, RELBLOCKS = NULL, phi = 10000)
}
\arguments{
\item{SX}{a vector of marker positions}

\item{CHRS}{a vector of marker chromosomes}

\item{RELBLOCKS}{a list of matrices containing, for each chromosome,
a matrix of starting and ending positions, and relevances}

\item{phi}{the range parameter to use for each marker, defaults to 10,000}
}
\description{
This function calculates the normalized spatial boost term in our prior
on theta, crossprod(w_j(phi), r), given a set of marker positions
and their chromosomes, as well as the starting and ending positions of
features of interest (e.g. genes) and their relevances.
}
\examples{
SX <- sort(runif(25, 0, 10))
CHRS <- rep(1, 25)
SITES <- list()
SITES[[1]] <- matrix(0, nrow = 5, ncol = 2)
SITES[[1]][1,] <- c(1, 3)
SITES[[1]][2,] <- c(2, 3)
SITES[[1]][3,] <- c(5, 6)
SITES[[1]][4,] <- c(7, 9)
SITES[[1]][5,] <- c(9.75, 10)
rel <- list()
rel[[1]] <- c(1, 1, 1, 1, 1)
RELBLOCKS <- list()
RELBLOCKS[[1]] <- relblock(SITES[[1]], rel[[1]])
spatialterm <- calculate_spatial_boost(SX,
                                       CHRS,
                                       RELBLOCKS,
                                       phi = 0.05)
plot(SX,
     spatialterm,
     xlab = "Marker position",
     ylab = "Spatial term",
     main = expression("Simple example with " * phi * " = .05"),
     pch = 20,
     col = "black",
     xlim = c(0, 10),
     ylim = c(-1, 1))
for(i in c(1,3,4,5)) {
  lines(x = SITES[[1]][i, 1:2],
        y = c(-.5, -.5),
        lwd = 2,
        col = "purple")
}
lines(x = SITES[[1]][2, 1:2],
      y = c(-.7, -.7),
      lwd = 2,
      col = "purple")
calculate_spatial_boost()
}
\keyword{boost}
\keyword{spatial}

