% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EM_beta_qual_delta}
\alias{EM_beta_qual_delta}
\title{EM M-step on beta utilizing SVD approximation}
\usage{
EM_beta_qual_delta(U, V, SIGMA, W, Z)
}
\arguments{
\item{U}{}

\item{V}{}

\item{SIGMA}{}

\item{W}{}

\item{Z}{}
}
\description{
This function performs part of the M-step of our EM algorithm
and computes the optimal value of beta utilizing a computational
speed-up by using the truncated SVD approximation to the design
matrix.
}
\examples{
EM_beta_qual_delta()
}
\keyword{Optimized}
\keyword{beta}
\keyword{of}
\keyword{value}

