% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EM_beta_qual_maximize}
\alias{EM_beta_qual_maximize}
\title{EM M-step on beta without utilizing SVD approximation}
\usage{
EM_beta_qual_maximize(Y, X, BETA, THETA, SIG2_0, KAPPA)
}
\arguments{
\item{Y}{the vector of response variables}

\item{X}{the full design matrix (additional covariates in front)}

\item{BETA}{the current value of beta}

\item{THETA}{the current expected values of theta}

\item{SIG2_0}{the current value of sigma squared}

\item{KAPPA}{the value of kappa in the prior on beta}
}
\description{
This function performs part of the M-step of our EM algorithm
and computes the optimal value of beta without utilizing any
computational speed-up.
}
\examples{
EM_beta_qual_maximize()
}
\keyword{Optimized}
\keyword{beta}
\keyword{of}
\keyword{value}

