% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EM_qual}
\alias{EM_qual}
\title{EM algorithm}
\usage{
EM_qual(X = NULL, Z = NULL, Y, weights = NULL, XI0 = -2, XI1 = 1,
  KAPPA = 1000, V = NULL, LAM = NULL, maxiter = 20, tol = 1e-04,
  print = TRUE, k = 0, SVD_X = NULL)
}
\arguments{
\item{X}{the matrix of genotypes}

\item{Z}{the matrix of additional covariates (optional)}

\item{Y}{the vector of response variables}

\item{weights}{the normalized sums of gene weights * relevances}

\item{XI0}{the hyper-parameter xi_0}

\item{XI1}{the hyper-parameter xi_1}

\item{KAPPA}{the value of kappa in the prior on beta}

\item{V}{a hyper-parameter in the prior on sigma squared (nu)}

\item{LAM}{a hyper-parameter in the prior on sigma squared (lambda)}

\item{maxiter}{maximum number of iterations for the EM algorithm}

\item{tol}{tolerance for convergence of the EM algorithm}

\item{print}{boolean flag for whether or not to print information}

\item{k}{number of additional covariates, defaults to 0}

\item{SVD_X}{list containing the SVD of cbind(Z, X)}
}
\description{
This function runs our EM algorithm once on a given GWAS data
set, assuming a binary trait for the response variable.
}
\examples{
EM_qual()
}
\keyword{EM}
\keyword{algorithm}

