% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EM_qual_filter}
\alias{EM_qual_filter}
\title{Expectation-Maximization filter}
\usage{
EM_qual_filter(X = NULL, Z = NULL, Y, weights = NULL, XI0 = -2,
  XI1 = 1, XI1_CONST = NULL, KAPPA = 1000, V = NULL, LAM = NULL,
  maxiter = 20, tol = 1e-04, print = TRUE, k = 0, SVD_X = NULL,
  per2rm = 0.25, desiredP = 100, bfdr_threshold = 0.05,
  ns_threshold = 10)
}
\arguments{
\item{X}{the matrix of genotypes}

\item{Z}{the matrix of additional covariates (optional)}

\item{Y}{the vector of response variables}

\item{weights}{the normalized sums of gene weights * relevances}

\item{XI0}{the hyper-parameter xi_0}

\item{XI1}{the hyper-parameter xi_1}

\item{XI1_CONST}{the desired ratio of XI0 / XI1 to keep}

\item{KAPPA}{a vector of values of kappa to use in the fitting}

\item{V}{a hyper-parameter in the prior on sigma squared (nu)}

\item{LAM}{a hyper-parameter in the prior on sigma squared (lambda)}

\item{maxiter}{the maximum number of iterations for the EM algorithm}

\item{tol}{the tolerance to achieve convergence for the EM algorithm}

\item{print}{a boolean flag for whether or not to print information}

\item{k}{the number of additional covariates, defaults to 0}

\item{SVD_X}{a list object containing the SVD approximation to cbind(Z, X)}

\item{per2rm}{percent of markers to remove at each iteration of the filter}

\item{desiredP}{the final desired number of markers}

\item{bfdr_threshold}{the threshold on EMBFDR}

\item{ns_threshold}{the threshold on the number of markers to be selected}
}
\description{
This function runs through the EM filter we propose and reduces
the number of markers in an initial data set to a desired number.
It returns a list containing the fitted model parameters and the
posterior predictive loss (PPL) at each iteration of the filter.
We recommend stopping the EM filter when the PPL is minimized.
}
\examples{
EM_qual_filter()
}
\keyword{EM}
\keyword{filter}

