% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EM_sigma_qual_maximize}
\alias{EM_sigma_qual_maximize}
\title{EM M-step on sigma squared}
\usage{
EM_sigma_qual_maximize(KAPPA, V, LAM, THETA, BETA)
}
\arguments{
\item{KAPPA}{the value of kappa in the prior on beta}

\item{V}{a hyper-parameter for the prior on sigma squared (nu)}

\item{LAM}{a hyper-parameter for the prior sigma squared (lambda)}

\item{THETA}{the current expected values of theta}

\item{BETA}{the current values of beta}
}
\description{
This function performs part of the M-step of our EM algorithm
and computes the optimal value of sigma squared
}
\examples{
EM_sigma_qual_maximize()
}
\keyword{Optimized}
\keyword{of}
\keyword{sigma}
\keyword{squared}
\keyword{value}

