% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genpg}
\alias{genpg}
\title{Polya-Gamma genpg function}
\usage{
genpg(n, Z)
}
\arguments{
\item{n}{an integer (first parameter of the distribution)}

\item{Z}{a floating point number (second parameter of the distribution)}
}
\description{
This function is one of several that is used to simulate a
Polya-Gamma random variable.
}
\examples{
genpg()
}
\keyword{genpg}

