% Generated by roxygen2 (4.0.2): do not edit by hand
\name{l}
\alias{l}
\title{l function}
\usage{
l(l1, l2)
}
\arguments{
\item{l1}{a floating point number (the first exponent in the sum)}

\item{l2}{a floating point number (the second exponent in the sum)}
}
\description{
This function is one of several that is used to compute
the log of a sum of exponents in a way that avoids overflow.
}
\examples{
l()
}
\keyword{l}

