% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lpnorm}
\alias{lpnorm}
\title{Polya-Gamma lpnorm function}
\usage{
lpnorm(z)
}
\arguments{
\item{z}{z-score}
}
\description{
This function returns the log of the area to the
left of the given value under the standard normal
probability distribution function.
}
\examples{
lpnorm()
}
\keyword{lpnorm}

