% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lse}
\alias{lse}
\title{lse function}
\usage{
lse(x)
}
\arguments{
\item{x}{a vector of exponents for which to compute log(sum(exp(x))).}
}
\description{
This function is one of several that is used to compute
the log of a sum of exponents in a way that avoids overflow.
}
\examples{
lse(c(0,0)) # log(exp(0) + exp(0)) = log(2) = 0.6931472
lse()
}
\keyword{lse}

