% Generated by roxygen2 (4.0.2): do not edit by hand
\name{par.omega.final}
\alias{par.omega.final}
\title{Polya-Gamma variable (omega) final sampler}
\usage{
par.omega.final(X, BETA, cl)
}
\arguments{
\item{X}{the design matrix}

\item{BETA}{the current values of beta}

\item{cl}{the cluster on which to run the sampler}
}
\description{
This function is one of two used to sample new values of
the polya-gamma latent variables (omega) in parallel.
}
\examples{
par.omega.final()
}
\keyword{polya-gamma}
\keyword{sampler}

