% Generated by roxygen2 (4.0.2): do not edit by hand
\name{par.omega.sampler}
\alias{par.omega.sampler}
\title{Polya-Gamma variable (omega) sampler}
\usage{
par.omega.sampler(INPUT)
}
\arguments{
\item{INPUT}{}
}
\description{
This function is one of two used to sample new values of
the polya-gamma latent variables (omega) in parallel.
}
\examples{
par.omega.sampler()
}
\keyword{polya-gamma}
\keyword{sampler}

