% Generated by roxygen2 (4.0.2): do not edit by hand
\name{phi_optimal}
\alias{phi_optimal}
\title{Range parameter (phi) selecting function}
\usage{
phi_optimal(X, SX, breakpoint = 30000, phigrid = c(10000, 25000, 50000,
  75000, 1e+05))
}
\arguments{
\item{X}{the matrix of genotypes}

\item{SX}{the vector of marker positions}

\item{breakpoint}{the minimum distance between adjacent blocks}

\item{phigrid}{a vector of possible values for phi}
}
\description{
This function breaks a chromosome into sub-regions and, for
each block, determines the optimal value of the range parameter,
phi, from a given set of values.
}
\examples{
phi_optimal()
}
\keyword{phi}

