% Generated by roxygen2 (4.0.2): do not edit by hand
\name{probexp}
\alias{probexp}
\title{Polya-Gamma probexp function}
\usage{
probexp(z, fz)
}
\arguments{
\item{z}{}

\item{fz}{}
}
\description{
This function is one of several that is used to simulate a
Polya-Gamma random variable.
}
\examples{
probexp()
}
\keyword{probexp}

