% Generated by roxygen2 (4.0.2): do not edit by hand
\name{relblock}
\alias{relblock}
\title{Relevance block calculation function}
\usage{
relblock(SITES, rel)
}
\arguments{
\item{SITES}{the matrix of starting and ending positions of features of interest}

\item{rel}{the vector of relevance values for each feature}
}
\description{
This function breaks a given set of (possibly) overlapping
features on a chromosome into unique blocks and averages the
relevance values for the overlapping regions.
}
\examples{
relblock()
}
\keyword{block}
\keyword{relevance}

