% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rpg}
\alias{rpg}
\title{Polya-Gamma rpg function}
\usage{
rpg(z, fz, pq)
}
\arguments{
\item{z}{}

\item{fz}{}

\item{pq}{}
}
\description{
This function is one of several that is used to simulate a
Polya-Gamma random variable.
}
\examples{
rpg()
}
\keyword{rpg}

