% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtigauss}
\alias{rtigauss}
\title{Polya-Gamma rtigauss function}
\usage{
rtigauss(z)
}
\arguments{
\item{z}{}
}
\description{
This function is one of several that is used to simulate a
Polya-Gamma random variable.
}
\examples{
rtigauss()
}
\keyword{rtigauss}

