% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SB_optimal_SVD}
\alias{SB_optimal_SVD}
\title{Truncated SVD approximation to X}
\usage{
SB_optimal_SVD(X, Z = NULL, MSEtol = 0.01)
}
\arguments{
\item{X}{a matrix of genotypes}

\item{Z}{a matrix of additional covariates, defaults to NULL}

\item{MSEtol}{the desired tolerance on the MSE, defaults to 1\%}
}
\description{
This function computes the truncated singular value
decomposition to a given set of genotypes, X, and a
given set of additional covariates, Z, for a desired
tolerance on the mean squared error.
}
\examples{
DATA <- simulate_gwas_data()
          SVDX <- SB_optimal_SVD(DATA$X)
SB_optimal_SVD()
}
\keyword{SVD}

