% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SB_qual_betasampler}
\alias{SB_qual_betasampler}
\title{Gibbs sampler for beta}
\usage{
SB_qual_betasampler(KAPPA, OMEGA, X, Y, THETA, SIG2_0)
}
\arguments{
\item{KAPPA}{the value of kappa in the prior on beta}

\item{OMEGA}{the current values of the latent polya gamma variables}

\item{X}{the design matrix}

\item{Y}{the response variables}

\item{THETA}{the current values of theta}

\item{SIG2_0}{the current values of sigma squared}
}
\description{
This function samples new values for beta from its
conditional posterior distribution given the current values
of the other parameters in the model.
}
\examples{
SB_qual_betasampler()
}
\keyword{beta}
\keyword{for}
\keyword{gibbs}
\keyword{sampler}

