% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SB_qual_complete_analysis}
\alias{SB_qual_complete_analysis}
\title{Complete Analysis Using the Spatial Boost Model}
\usage{
SB_qual_complete_analysis(X = NULL, Z = NULL, Y = NULL, weights = NULL,
  EMXI0 = -2, EMXI1 = 0, XI1_CONST = NULL, GSXI0 = -2, GSXI1 = 0,
  KAPPA = 10000, V = 3, LAM = 0.1, maxiter = 100, tol = 1e-04,
  print = TRUE, k = 0, SVD_X = NULL, per2rm = 0.25, desiredP = 100,
  bfdr_threshold = 0.05, ns_threshold = 10, NCLUST = 1, NSAMPS = 1000,
  FINDEX = NULL)
}
\arguments{
\item{X}{the matrix of genotypes}

\item{Z}{the matrix of additional covariates (optional)}

\item{Y}{the vector of response variables}

\item{weights}{the normalized sums of gene weights * relevances}

\item{XI1_CONST}{the desired ratio of XI0 / XI1 to keep}

\item{KAPPA}{a vector of values of kappa to use in the fitting}

\item{V}{a hyper-parameter in the prior on sigma squared (nu)}

\item{LAM}{a hyper-parameter in the prior on sigma squared (lambda)}

\item{maxiter}{the maximum number of iterations for the EM algorithm}

\item{tol}{the tolerance to achieve convergence for the EM algorithm}

\item{print}{a boolean flag for whether or not to print information}

\item{k}{the number of additional covariates, defaults to 0}

\item{SVD_X}{a list object containing the SVD approximation to cbind(Z, X)}

\item{per2rm}{percent of markers to remove at each iteration of the filter}

\item{desiredP}{the final desired number of markers}

\item{bfdr_threshold}{the threshold on EMBFDR}

\item{ns_threshold}{the threshold on the number of markers to be selected}

\item{NCLUST}{the number of clusters to use in the parallel sampling of omega}

\item{NSAMPS}{the desired number of samples}

\item{FINDEX}{the index of the EM filter to use}

\item{XI0}{the hyper-parameter xi_0}

\item{XI1}{the hyper-parameter xi_1}
}
\description{
This function runs through the EM filter we propose and reduces
the number of markers in an initial data set to a desired number
and then runs the Gibbs sampler on the final set of retained markers.
It returns a list containing the fitted model parameters and the
posterior predictive loss (PPL) at each iteration of the filter,
as well as a list containing the samples drawn from the gibbs sampler.
}
\examples{
SB_qual_complete_analysis()
}
\keyword{EM}
\keyword{filter}

