% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SB_qual_gibbssampler}
\alias{SB_qual_gibbssampler}
\title{Gibbs sampler}
\usage{
SB_qual_gibbssampler(GX, Y, weights = NULL, SIG2 = 0.01, KAPPA = 100,
  NSAMPS = 1000, NCLUST = 1, XI0 = -2, XI1 = 0, V = NULL,
  LAM = NULL, k = 0, print = TRUE)
}
\arguments{
\item{GX}{the design matrix}

\item{Y}{the response variables}

\item{weights}{the normalized sums of the gene weights * relevances}

\item{SIG2}{the initial value of sigma squared}

\item{KAPPA}{the value of kappa in the prior on beta}

\item{NSAMPS}{the desired number of samples}

\item{NCLUST}{the number of clusters to use in the parallel sampling}

\item{XI0}{the hyper-parameter xi_0}

\item{XI1}{the hyper-parameter xi_1}

\item{V}{a hyper-parameter in the prior on sigma squared (nu)}

\item{LAM}{a hyper-parameter in the prior on sigma squared (lambda)}

\item{k}{the number of additional covariates, defaults to 0}

\item{print}{boolean flag for whether or not to print information}
}
\description{
This function runs our overall gibbs sampler for the
spatial boost model making use of the data-augmentation
strategy that exploits polya-gamma latent variables
}
\examples{
SB_qual_gibbssampler()
}
\keyword{gibbs}
\keyword{sampler}

