% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SB_qual_sigma2sampler}
\alias{SB_qual_sigma2sampler}
\title{Gibbs sampler for sigma squared}
\usage{
SB_qual_sigma2sampler(KAPPA, V, LAM, THETA, BETA)
}
\arguments{
\item{KAPPA}{the value of kappa in the prior on beta}

\item{V}{a hyper-parameter for the prior on sigma squared (nu)}

\item{LAM}{a hyper-parameter for the prior on sigma squared (lambda)}

\item{THETA}{the current values of theta}

\item{BETA}{the current values of beta}
}
\description{
This function samples new values for sigma squared from its
conditional posterior distribution given the current values
of the other parameters in the model.
}
\examples{
SB_qual_sigma2sampler()
}
\keyword{for}
\keyword{gibbs}
\keyword{sampler}
\keyword{sigma}
\keyword{squared}

