% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SB_qual_thetasampler}
\alias{SB_qual_thetasampler}
\title{Gibbs sampler for theta}
\usage{
SB_qual_thetasampler(KAPPA, GENEBOOST, BETA, SIG2_0)
}
\arguments{
\item{KAPPA}{the value of kappa in the prior on beta}

\item{GENEBOOST}{the negative of the gene boost term for each marker}

\item{BETA}{the current value of beta}

\item{SIG2_0}{the current value of sigma squared}
}
\description{
This function samples new values for the thetas from their
conditional posterior distributions given the current values
of the other parameters in the model.
}
\examples{
SB_qual_thetasampler()
}
\keyword{for}
\keyword{gibbs}
\keyword{sampler}
\keyword{theta}

