% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulate_gwas_data}
\alias{simulate_gwas_data}
\title{Simulate GWAS Data function}
\usage{
simulate_gwas_data(n = 50, p = 10, alpha = 0.2, KAPPA = 100, sig2 = 1)
}
\arguments{
\item{n}{sample size, defaults to 50}

\item{p}{number of markers, defaults to 10}

\item{alpha}{the prior probability of association, defaults to 0.2}

\item{KAPPA}{value of kappa in the prior on beta, defaults to 100}

\item{sig2}{value of sigma squared in the prior on beta, defaults to 1}
}
\description{
This function can be used to simulate a simple
GWAS data set with a binary trait response variable.
}
\examples{
DATA <- simulate_gwas_data(n = 100,
                                     p = 1000,
                                     alpha = 0.01,
                                     KAPPA = 1e6,
                                     sig2 = 1e-4)
          which(DATA$THETA == 1) # list of causal markers
          summary(colMeans(DATA$X) / 2) # summary of minor allele frequencies
simulate_gwas_data()
}
\keyword{data}
\keyword{simulate}

